services/celery_worker/requirements.txt
celery[redis]==5.3.4
redis==4.6.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
httpx==0.25.2
google-api-python-client==2.108.0
google-auth==2.25.2
twilio==9.2.3

